+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryUserStorage.java
@@ -15,8 +15,8 @@ import java.util.Map;
 @Component
 public class InMemoryUserStorage implements UserStorage {
 
-    private final Map<Long, User> users = new HashMap<>();
-    private long counter = 0L;
+    private final Map<Integer, User> users = new HashMap<>();
+    private int counter = 0;
 
     @Override
     public List<User> getUsers() {
@@ -25,7 +25,7 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User add(User user) {
-        final long id = ++counter;
+        final int id = ++counter;
         user.setId(id);
         users.put(user.getId(), user);
         return user;
